Clean code, Add Packages
- move linux banner and pslist processing to `volatility/linux.go`
- move windows pslist processing to `volatility/windows.go`
- add helper utils
- add user STIX objects where available
