+++ forensix/go.mod
@@ -7,6 +7,4 @@ require github.com/TcM1911/stix2 v0.8.0
 require (
 	github.com/google/uuid v1.3.0 // indirect
 	github.com/ugorji/go/codec v1.2.6 // indirect
-	golang.org/x/lint v0.0.0-20210508222113-6edffad5e616 // indirect
-	golang.org/x/tools v0.0.0-20200130002326-2f3ba24bd6e7 // indirect
 )
