+++ forensix/utils/stix.go
@@ -1,0 +1,8 @@
+package utils
+
+import (
+	"github.com/TcM1911/stix2"
+)
+
+//Collection is a new STIX 2.1 collection for holding STIX objects
+var Collection = stix2.New()
